requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
youtube-transcript-api>=0.6.2
google-genai>=0.2.0
chromadb>=0.5.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
